% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-low-level.R
\name{add_data}
\alias{add_data}
\title{Add Data to a Repository}
\usage{
add_data(rdf_data, access_options, data_format = "application/x-trig")
}
\arguments{
\item{rdf_data}{character. RDF data containing the triples to be submitted.}

\item{access_options}{output of \code{basic_triplestore_access} or \code{api_triplestore_access}}

\item{data_format}{currently only `application/x-trig` is supported. A variant of Turtle.}
}
\value{
the response of the `httr::POST` against the endpoint
}
\description{
Add Data to a Repository
}
\examples{
rdf_data = c("<http://test.net/mytest1> {
  <http://test.net/mysubject2> <http://test.net/has-value> '5' .
}")
add_data(rdf_data, access_options = graphdb)

}
