% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-high-level.R
\name{add_data_factory}
\alias{add_data_factory}
\title{Add Data Function Factory}
\usage{
add_data_factory(access_options, prefixes)
}
\arguments{
\item{prefixes}{}
}
\description{
Wraps \code{add_data} to simply submit a Turtle/Trig file to a triplestore
}
