% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-high-level.R
\name{add_trig_file_to_graphdb}
\alias{add_trig_file_to_graphdb}
\title{The function `add_trig_file_to_graphdb` retrieves the rdf data from a trig file and inserts each line into an element of a list.
Then, this list is imported to graphdb triplestore using rdfr4::add_factory() functtion.}
\usage{
add_trig_file_to_graphdb(graphdb_access_options, prefixes, trig_file)
}
\arguments{
\item{graphdb_access_options}{graphdb triplestore access options.}

\item{prefixes}{named character}

\item{trig_file}{file name of the rdf format containing context (named graph).}
}
\description{
The function `add_trig_file_to_graphdb` retrieves the rdf data from a trig file and inserts each line into an element of a list.
Then, this list is imported to graphdb triplestore using rdfr4::add_factory() functtion.
}
