% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{AnonRDF}
\alias{AnonRDF}
\title{Create a list of RDF statements that all share the same blank subject node}
\description{
Create a list of RDF statements that all share the same blank subject node

Create a list of RDF statements that all share the same blank subject node
}
\seealso{
Other anonymous rdf: 
\code{\link{is.AnonRDF}()}
}
\concept{anonymous rdf}
\section{Super class}{
\code{\link[rdf4r:ResourceDescriptionFramework]{rdf4r::ResourceDescriptionFramework}} -> \code{anonymous_rdf}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-add_triple}{\code{AnonRDF$add_triple()}}
\item \href{#method-add_triples}{\code{AnonRDF$add_triples()}}
\item \href{#method-serialize}{\code{AnonRDF$serialize()}}
\item \href{#method-clone}{\code{AnonRDF$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rdf4r" data-topic="ResourceDescriptionFramework" data-id="add_triples_extended">}\href{../../rdf4r/html/ResourceDescriptionFramework.html#method-add_triples_extended}{\code{rdf4r::ResourceDescriptionFramework$add_triples_extended()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdf4r" data-topic="ResourceDescriptionFramework" data-id="get_list">}\href{../../rdf4r/html/ResourceDescriptionFramework.html#method-get_list}{\code{rdf4r::ResourceDescriptionFramework$get_list()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdf4r" data-topic="ResourceDescriptionFramework" data-id="get_prefixes">}\href{../../rdf4r/html/ResourceDescriptionFramework.html#method-get_prefixes}{\code{rdf4r::ResourceDescriptionFramework$get_prefixes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdf4r" data-topic="ResourceDescriptionFramework" data-id="initialize">}\href{../../rdf4r/html/ResourceDescriptionFramework.html#method-initialize}{\code{rdf4r::ResourceDescriptionFramework$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdf4r" data-topic="ResourceDescriptionFramework" data-id="serialize_to_file">}\href{../../rdf4r/html/ResourceDescriptionFramework.html#method-serialize_to_file}{\code{rdf4r::ResourceDescriptionFramework$serialize_to_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdf4r" data-topic="ResourceDescriptionFramework" data-id="set_context">}\href{../../rdf4r/html/ResourceDescriptionFramework.html#method-set_context}{\code{rdf4r::ResourceDescriptionFramework$set_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdf4r" data-topic="ResourceDescriptionFramework" data-id="set_list">}\href{../../rdf4r/html/ResourceDescriptionFramework.html#method-set_list}{\code{rdf4r::ResourceDescriptionFramework$set_list()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_triple"></a>}}
\if{latex}{\out{\hypertarget{method-add_triple}{}}}
\subsection{Method \code{add_triple()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnonRDF$add_triple(predicate, object)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_triples"></a>}}
\if{latex}{\out{\hypertarget{method-add_triples}{}}}
\subsection{Method \code{add_triples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnonRDF$add_triples(ll)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-serialize"></a>}}
\if{latex}{\out{\hypertarget{method-serialize}{}}}
\subsection{Method \code{serialize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnonRDF$serialize(context)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnonRDF$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
