% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-low-level.R
\name{api_triplestore_access}
\alias{api_triplestore_access}
\title{API Triplestore Access}
\usage{
api_triplestore_access(server_url, repository, user = as.character(NA),
  password = as.character(NA))
}
\arguments{
\item{server_url}{character. The URL of the triplestore.}

\item{repository}{character. The ID of the repository to which you want to connect to.}

\item{api_key}{a string, the API key used for API-style authentication.}

\item{secret}{a string, the secret string corresponding to the API key
needed for API-style authentication.}
}
\value{
list. Contains the server access options.
}
\description{
TODO: UNFINISHED
}
\details{
Create an object with the access options for a triple-store. If successful it also outputs to the screen the protocol version. If unsuccessful it raises an error. Uses API authentication.
}
\examples{

}
\seealso{
\code{api_triplestore_access}
}
