% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{blank_node}
\alias{blank_node}
\title{A blank node}
\format{
An object of class \code{identifier} of length 4.
}
\usage{
blank_node
}
\description{
A blank node
}
\seealso{
Other semantic elements: 
\code{\link{rdf_type}},
\code{\link{semantic_elements}},
\code{\link{xsd_boolean}},
\code{\link{xsd_date}},
\code{\link{xsd_double}},
\code{\link{xsd_integer}},
\code{\link{xsd_string}}
}
\concept{semantic elements}
\keyword{datasets}
