% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-vector.R
\name{DynVector}
\alias{DynVector}
\title{An Amortized Vector}
\description{
This object stores a preallocated list. You can add objects to the list
with \code{add} and add lists to the list with \code{add_list}. When the
objects to be added exceed the dimensions of the list, the list is re-
allocated with double the size. This amortizes the runtime if you keep
growing the dynamic vector.
}
\details{
Use this object if you need an object with a mutable states that you
keep growing. An example with be a list of RDF triples.
}
\examples{

v = DynVector$new(3)
v$add("test")
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DynVector$new()}}
\item \href{#method-add}{\code{DynVector$add()}}
\item \href{#method-add_list}{\code{DynVector$add_list()}}
\item \href{#method-get}{\code{DynVector$get()}}
\item \href{#method-clone}{\code{DynVector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynVector$new(size)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynVector$add(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_list"></a>}}
\if{latex}{\out{\hypertarget{method-add_list}{}}}
\subsection{Method \code{add_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynVector$add_list(l)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynVector$get()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynVector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
