% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier.R
\name{fidentifier}
\alias{fidentifier}
\title{Identifier Constructor via a List of Lookup Functions}
\usage{
fidentifier(fun, ..., prefixes)
}
\arguments{
\item{fun}{\code{list} of lookup functions to be tried. The functions
should have the same arguments (see \code{...} below). As soon as we have
a unique match the function execution halts. If there is no match, a
URI with the base prefix (the one indiciated by "_base") and a UUID
will be generated.}

\item{...}{arguments to be passed to every lookup function. The arguments
must be "representable", i.e. either a literal or a identifier}

\item{prefixes}{Named \code{character} that contains the prefixes to be
used for identifier construction after each lookup.}
}
\description{
Identifier Constructor via a List of Lookup Functions
}
