% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-low-level.R
\name{get_protocol_version}
\alias{get_protocol_version}
\title{Get Protocol Version}
\usage{
get_protocol_version(access_options)
}
\arguments{
\item{access_options}{list containing the graph database connectivity options,                returned by the helper function \code{basic_triplestore_access} or \code{api_triplestore_access}}
}
\value{
integer containing the protocol version if connectivity is OK.
}
\description{
Test connectivity to the graph database and get the communication protocol
version.
}
\details{
This function tests the connectivity to the graph database. If there is connectivity it will return the protocol version as an integer. If there is no connectivity, an error will be raised.
}
\examples{
get_protocol_version(graphdb)

}
\seealso{
\code{basic_triplestore_access} in order to see how to create the connection options for a triple store with basic HTTP authentication and \code{api_access_options} for API-based authentication to a triplestore
}
