% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier.R
\name{identifier}
\alias{identifier}
\title{Construct Identifier from ID}
\usage{
identifier(id, prefix)
}
\arguments{
\item{id}{character(1). The part of identifier after the prefix.}

\item{prefix}{named character(1). The prefix. See Examples.}
}
\value{
identifier object (a type of list).
}
\description{
Construct Identifier from ID
}
\examples{
sample_id = identifier("57d68e07-8315-4b30-9a8e-57226fd815d7",
  prefix = c(openbiodiv = "http://openbiodiv.net/"))
}
