% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier.R
\name{identifier_factory}
\alias{identifier_factory}
\title{Manufacturing Identifier Constructors}
\usage{
identifier_factory(..., FUN = list(...), prefixes, def_prefix)
}
\arguments{
\item{...}{arguments to be passed to every lookup function.}

\item{prefixes}{Named \code{character} that contains the prefixes to be
used for identifier construction after each lookup.}

\item{def_prefix}{named character of length 1. default prefix The prefix
to be used if lookup fails.}
}
\value{
an identifier constructor function with one parameter
  (\code{label})
}
\description{
Manufacturing Identifier Constructors
}
\examples{

openbiodiv_id = identifier_factory(simple_lookup,
  prefixes = prefixes,
  def_prefix = c(openbiodiv = "http://openbiodiv.net/"))

openbiodiv_id("Teodor Georgiev")
openbiodiv_id("Pavel Stoev")
}
