% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier.R
\name{is.identifier}
\alias{is.identifier}
\title{Is the object an identifier?}
\usage{
is.identifier(x)
}
\arguments{
\item{x}{object to check}
}
\value{
logical
}
\description{
Is the object an identifier?
}
\seealso{
Other identifier functions: 
\code{\link{identifier}()},
\code{\link{print.identifier}()},
\code{\link{represent.identifier}()}
}
\concept{identifier functions}
