% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literal.R
\name{is.literal}
\alias{is.literal}
\title{Is the object a literal?}
\usage{
is.literal(x)
}
\arguments{
\item{x}{object to check}
}
\value{
logical
}
\description{
Is the object a literal?
}
\seealso{
Other literal functions: \code{\link{literal}},
  \code{\link{print.literal}},
  \code{\link{represent.literal}}
}
