% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-low-level.R
\name{list_repositories}
\alias{list_repositories}
\title{List Repositories}
\usage{
list_repositories(access_options)
}
\arguments{
\item{access_options}{object returned from \code{basic_triplestore_access} or
\code{api_triplestore_access}}
}
\value{
data.frame. Contains repository information for that access point.
}
\description{
List Repositories
}
\examples{
list_repositories(graphdb)

}
