% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literal.R
\name{literal}
\alias{literal}
\title{Literal Construction}
\usage{
literal(text_value, xsd_type, lang)
}
\arguments{
\item{text_value}{\code{character} The textual value of literal.}

\item{xsd_type}{\code{identifier} The XSD type.}

\item{lang}{\code{character} The language code.}
}
\value{
a \code{literal} object, which is a \code{list} with the
  following fields:\itemize{
  \item{\code{text_value} \code{character} The textual value of literal.}
  \item{\code{xsd_type} \code{identifier} The XSD type.}
  \item{\code{lang} \code{character} The language code.}
  \item{\code{squote} \code{character} The value with added quotes and
   types,  e.g. "'John'^^xsd:string"}}
}
\description{
\code{literal} constructs a \code{literal} object.
}
\details{
If for some reason the supplied text value is meaningless (empty string,
just spaces, invalid type, etc.), the returned object is the equivalent
of NA
}
\examples{
# see vignette
}
