% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier.R
\name{lookup_identifier}
\alias{lookup_identifier}
\title{Construct Identifier via Lookup}
\usage{
lookup_identifier(label, prefixes, FUN = list(...), ...)
}
\arguments{
\item{label}{character(1). Parameter that will be passed to the lookup
functions. See \code{...} for details.}

\item{prefixes}{named character. Contains the prefixes.}

\item{FUN}{list of \code{...}}

\item{...}{(lookup) functions that will be executed in order to
obtain the identifier. The functions should have one argument to which
\code{label} will be assigned during the call. As soon as we have a
unique match the function execution halts. If there is no match, a
URI with the base prefix (the one indiciated by "_base") and a UUID
will be generated.}
}
\value{
identifier object (a type of list)
}
\description{
Construct Identifier via Lookup
}
