% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{pasteif}
\alias{pasteif}
\title{Paste If}
\usage{
pasteif(..., cond, return_value = NULL)
}
\arguments{
\item{...}{one or more R objects, to be passed to paste0 if cond is TRUE.}

\item{cond}{boolean, condition to be true.}

\item{return_value}{}
}
\value{
string. If condition is false returns \code{return_value}. default is NULL
}
\description{
Pastes something only if the condition is met. Uses paste0.
}
\examples{
pasteif("1", "st", cond = (3 < 2))

}
