% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix.R
\name{prefix_serializer}
\alias{prefix_serializer}
\title{Serializes a Prefix Vector}
\usage{
prefix_serializer(prefixes, reqd = names(prefixes), lang = "SPARQL")
}
\arguments{
\item{prefixes}{named character. Contains prefixes. Names are the prefixes.}

\item{reqd}{a character vector of needed prefixes, can be missing, then take all.}

\item{lang}{character. The serialization language. One of \code{"SPARQL"} or
\code{"Turtle"}. Default is \code{"SPARQL"}}
}
\value{
prefix serialization.
}
\description{
If \code{reqd} is left missing, all prefixes will be returned.
The individual prefixes may or may not end with ":". If they don't,
it will be added during function execution.
}
\details{
Note the base prefix should be denoted as "_base". Its seriazliation is then

\code{PREFIX : partial_uri}
}
\examples{
prefixes = c(rdfs = "<http://www.w3.org/2000/01/rdf-schema#>", foaf = "<http://xmlns.com/foaf/0.1/>", openbiodiv = "<http://openbiodiv.net/>")
prefix_serializer(prefixes, reqd = c("rdfs", "openbiodiv"))
prefix_serializer(prefixes, reqd = c("rdfs", "openbiodiv"), lang = "Turtle")

}
