% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier.R
\name{print.identifier}
\alias{print.identifier}
\title{Outputs an identifier in a default way}
\usage{
\method{print}{identifier}(id)
}
\arguments{
\item{id}{\code{identifier}}
}
\value{
\code{character} default representation.
}
\description{
Outputs an identifier in a default way
}
\seealso{
Other identifier functions: 
\code{\link{identifier}()},
\code{\link{is.identifier}()},
\code{\link{represent.identifier}()}
}
\concept{identifier functions}
