% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literal.R
\name{print.literal}
\alias{print.literal}
\title{Outputs a literal in a default way}
\usage{
\method{print}{literal}(ll)
}
\arguments{
\item{ll}{\code{literal}}
}
\value{
\code{character} default representation.
}
\description{
Outputs a literal in a default way
}
\seealso{
Other literal functions: \code{\link{is.literal}},
  \code{\link{literal}}, \code{\link{represent.literal}}
}
