% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{rdf_type}
\alias{rdf_type}
\title{RDF Type}
\format{An object of class \code{identifier} of length 4.}
\usage{
rdf_type
}
\description{
RDF Type
}
\seealso{
Other semantic elements: \code{\link{blank_node}},
  \code{\link{semantic_elements}}, \code{\link{xsd_date}},
  \code{\link{xsd_integer}}, \code{\link{xsd_string}},
  \code{\link{xsd_uri}}
}
\concept{semantic elements}
\keyword{datasets}
