% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{RdfLibBackend}
\alias{RdfLibBackend}
\title{RDF Class with \code{rdflib} backend}
\description{
RDF Class with \code{rdflib} backend

RDF Class with \code{rdflib} backend
}
\section{Super class}{
\code{\link[rdf4r:ResourceDescriptionFramework]{rdf4r::ResourceDescriptionFramework}} -> \code{RdfLibBackend}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RdfLibBackend$new()}}
\item \href{#method-nquad}{\code{RdfLibBackend$nquad()}}
\item \href{#method-clone}{\code{RdfLibBackend$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rdf4r" data-topic="ResourceDescriptionFramework" data-id="add_triple">}\href{../../rdf4r/html/ResourceDescriptionFramework.html#method-add_triple}{\code{rdf4r::ResourceDescriptionFramework$add_triple()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdf4r" data-topic="ResourceDescriptionFramework" data-id="add_triples">}\href{../../rdf4r/html/ResourceDescriptionFramework.html#method-add_triples}{\code{rdf4r::ResourceDescriptionFramework$add_triples()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdf4r" data-topic="ResourceDescriptionFramework" data-id="add_triples_extended">}\href{../../rdf4r/html/ResourceDescriptionFramework.html#method-add_triples_extended}{\code{rdf4r::ResourceDescriptionFramework$add_triples_extended()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdf4r" data-topic="ResourceDescriptionFramework" data-id="get_list">}\href{../../rdf4r/html/ResourceDescriptionFramework.html#method-get_list}{\code{rdf4r::ResourceDescriptionFramework$get_list()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdf4r" data-topic="ResourceDescriptionFramework" data-id="get_prefixes">}\href{../../rdf4r/html/ResourceDescriptionFramework.html#method-get_prefixes}{\code{rdf4r::ResourceDescriptionFramework$get_prefixes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdf4r" data-topic="ResourceDescriptionFramework" data-id="ntriples">}\href{../../rdf4r/html/ResourceDescriptionFramework.html#method-ntriples}{\code{rdf4r::ResourceDescriptionFramework$ntriples()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdf4r" data-topic="ResourceDescriptionFramework" data-id="serialize">}\href{../../rdf4r/html/ResourceDescriptionFramework.html#method-serialize}{\code{rdf4r::ResourceDescriptionFramework$serialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdf4r" data-topic="ResourceDescriptionFramework" data-id="serialize_to_file">}\href{../../rdf4r/html/ResourceDescriptionFramework.html#method-serialize_to_file}{\code{rdf4r::ResourceDescriptionFramework$serialize_to_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdf4r" data-topic="ResourceDescriptionFramework" data-id="set_context">}\href{../../rdf4r/html/ResourceDescriptionFramework.html#method-set_context}{\code{rdf4r::ResourceDescriptionFramework$set_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdf4r" data-topic="ResourceDescriptionFramework" data-id="set_list">}\href{../../rdf4r/html/ResourceDescriptionFramework.html#method-set_list}{\code{rdf4r::ResourceDescriptionFramework$set_list()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdfLibBackend$new(context)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nquad"></a>}}
\if{latex}{\out{\hypertarget{method-nquad}{}}}
\subsection{Method \code{nquad()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdfLibBackend$nquad(subject, predicate, object, context)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdfLibBackend$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
