% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{rdfs_label}
\alias{rdfs_label}
\title{Has Label Property}
\format{An object of class \code{identifier} of length 4.}
\usage{
rdfs_label
}
\description{
Has Label Property
}
\seealso{
Other semantic elements: \code{\link{blank_node}},
  \code{\link{rdf_type}}, \code{\link{semantic_elements}},
  \code{\link{xsd_date}}, \code{\link{xsd_integer}},
  \code{\link{xsd_string}}, \code{\link{xsd_uri}}
}
\concept{semantic elements}
\keyword{datasets}
