% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier.R
\name{represent.identifier}
\alias{represent.identifier}
\title{Outputs an identifier in a default way (not print)}
\usage{
\method{represent}{identifier}(id)
}
\arguments{
\item{id}{\code{identifier}}
}
\value{
\code{character} default representation.
}
\description{
Outputs an identifier in a default way (not print)
}
