% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literal.R
\name{represent.literal}
\alias{represent.literal}
\title{Outputs a literal in a default way (not print!)}
\usage{
\method{represent}{literal}(ll)
}
\arguments{
\item{ll}{\code{literal}}
}
\value{
\code{character} default representation.
}
\description{
Outputs a literal in a default way (not print!)
}
\seealso{
Other literal functions: \code{\link{is.literal}},
  \code{\link{literal}}, \code{\link{print.literal}}

Other representables: \code{\link{represent.identifier}},
  \code{\link{represent}}
}
