% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{represent}
\alias{represent}
\title{Defualt representation}
\usage{
represent(x)
}
\arguments{
\item{x}{}
}
\description{
Defualt representation
}
\seealso{
Other representables: \code{\link{represent.identifier}},
  \code{\link{represent.literal}}
}
\concept{representables}
