% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{ResourceDescriptionFramework}
\alias{ResourceDescriptionFramework}
\title{Mutable RDF Object}
\description{
This is a mutable object. Adds, stores, and serializes RDF statements
efficiently. Uses an amortized vector \code{DynVector} as a internal
storage. You can inherit this class and override to create your own
implementation and still retain syntax compatibility.
}
\seealso{
Other rdf: 
\code{\link{is.ResourceDescriptionFramework}()}
}
\concept{rdf}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ResourceDescriptionFramework$new()}}
\item \href{#method-set_context}{\code{ResourceDescriptionFramework$set_context()}}
\item \href{#method-get_prefixes}{\code{ResourceDescriptionFramework$get_prefixes()}}
\item \href{#method-add_triple}{\code{ResourceDescriptionFramework$add_triple()}}
\item \href{#method-add_triples}{\code{ResourceDescriptionFramework$add_triples()}}
\item \href{#method-get_list}{\code{ResourceDescriptionFramework$get_list()}}
\item \href{#method-set_list}{\code{ResourceDescriptionFramework$set_list()}}
\item \href{#method-serialize}{\code{ResourceDescriptionFramework$serialize()}}
\item \href{#method-serialize_to_file}{\code{ResourceDescriptionFramework$serialize_to_file()}}
\item \href{#method-add_triples_extended}{\code{ResourceDescriptionFramework$add_triples_extended()}}
\item \href{#method-ntriples}{\code{ResourceDescriptionFramework$ntriples()}}
\item \href{#method-clone}{\code{ResourceDescriptionFramework$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceDescriptionFramework$new(size = 1e+06)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_context"></a>}}
\if{latex}{\out{\hypertarget{method-set_context}{}}}
\subsection{Method \code{set_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceDescriptionFramework$set_context(context)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{context}}{\code{identifier}. The named graph of where the statements
are stored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_prefixes"></a>}}
\if{latex}{\out{\hypertarget{method-get_prefixes}{}}}
\subsection{Method \code{get_prefixes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceDescriptionFramework$get_prefixes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_triple"></a>}}
\if{latex}{\out{\hypertarget{method-add_triple}{}}}
\subsection{Method \code{add_triple()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceDescriptionFramework$add_triple(subject, predicate, object)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_triples"></a>}}
\if{latex}{\out{\hypertarget{method-add_triples}{}}}
\subsection{Method \code{add_triples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceDescriptionFramework$add_triples(ll)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_list"></a>}}
\if{latex}{\out{\hypertarget{method-get_list}{}}}
\subsection{Method \code{get_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceDescriptionFramework$get_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_list"></a>}}
\if{latex}{\out{\hypertarget{method-set_list}{}}}
\subsection{Method \code{set_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceDescriptionFramework$set_list(triple_vector)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-serialize"></a>}}
\if{latex}{\out{\hypertarget{method-serialize}{}}}
\subsection{Method \code{serialize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceDescriptionFramework$serialize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-serialize_to_file"></a>}}
\if{latex}{\out{\hypertarget{method-serialize_to_file}{}}}
\subsection{Method \code{serialize_to_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceDescriptionFramework$serialize_to_file(file_name, progress_bar = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_triples_extended"></a>}}
\if{latex}{\out{\hypertarget{method-add_triples_extended}{}}}
\subsection{Method \code{add_triples_extended()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceDescriptionFramework$add_triples_extended(
  data,
  subject_column_label = "",
  subject_column_name = "",
  subject_rdf_prefix = "",
  predicate,
  object_column_label = "",
  object_column_name = "",
  object_rdf_prefix = "",
  progress_bar = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ntriples"></a>}}
\if{latex}{\out{\hypertarget{method-ntriples}{}}}
\subsection{Method \code{ntriples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceDescriptionFramework$ntriples()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceDescriptionFramework$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
