% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\name{semantic_elements}
\alias{semantic_elements}
\title{Predifined resource identifiers from widely-spread ontologies}
\description{
Predifined resource identifiers from widely-spread ontologies
}
\seealso{
Other semantic elements: 
\code{\link{blank_node}},
\code{\link{rdf_type}},
\code{\link{xsd_boolean}},
\code{\link{xsd_date}},
\code{\link{xsd_double}},
\code{\link{xsd_integer}},
\code{\link{xsd_string}}
}
\concept{semantic elements}
