% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{strip_angle}
\alias{strip_angle}
\title{Strips (or Adds) Angular Brackets from a URI if Present}
\usage{
strip_angle(partial_uri, reverse = FALSE)
}
\arguments{
\item{partial_uri}{the URI to strip angular brackets from}

\item{reverse}{FALSE, if true will put angular brakcets instead}
}
\value{
URI with stripped angular brackets around it
}
\description{
Strips (or Adds) Angular Brackets from a URI if Present
}
