% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{strip_trailing_symbol}
\alias{strip_trailing_symbol}
\title{Strip Trailing Slash or Colon If Present}
\usage{
strip_trailing_symbol(x)
}
\arguments{
\item{x}{character vector.}
}
\value{
x without the trailing slash, if present
}
\description{
Symbols that are stripped if present: /, :
}
\details{
Works only on UNIX.
}
\examples{
strip_trailing_symbol("/media/obkms/plazi-corpus-xml/")
strip_trailing_symbol("rdfs:")
strip_trailing_symbol("rdfs")


}
