% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-low-level.R
\name{submit_sparql_update}
\alias{submit_sparql_update}
\title{Submit a SPARQL Query to a Triplestore (UPDATE)}
\usage{
submit_sparql_update(query, access_options)
}
\arguments{
\item{query}{\code{character} the SPARQL UPDATE query. See Details.}

\item{access_options}{\code{triplestore_access_options} object.}
}
\value{

}
\description{
This submits an UPDATE (in the sense of the CRUD methodology) to a
SPARQL endpoint. For READ queries see \code{submit_sparql}.
}
\details{
Here are the SPARQL operations that are supported:

LOAD
CLEAR
DROP
ADD
MOVE
COPY
CREATE
INSERT
DELETE
}
\examples{
\dontrun{
#' drop_query = "DROP GRAPH <http::/openbiodiv.net/test123>"

submit_sparql_update(drop_query, access_options = graphdb)
}

}
\seealso{
\code{submit_sparql}
}
