% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{xsd_boolean}
\alias{xsd_boolean}
\title{XSD Boolean}
\format{
An object of class \code{identifier} of length 4.
}
\usage{
xsd_boolean
}
\description{
XSD Boolean
}
\seealso{
Other semantic elements: 
\code{\link{blank_node}},
\code{\link{rdf_type}},
\code{\link{semantic_elements}},
\code{\link{xsd_date}},
\code{\link{xsd_double}},
\code{\link{xsd_integer}},
\code{\link{xsd_string}}
}
\concept{semantic elements}
\keyword{datasets}
