% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{xsd_integer}
\alias{xsd_integer}
\title{XSD Integer Type}
\format{An object of class \code{identifier} of length 4.}
\usage{
xsd_integer
}
\description{
XSD Integer Type
}
\keyword{datasets}
