% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{xsd_string}
\alias{xsd_string}
\alias{xsd_string}
\alias{rdfs_label}
\title{XSD String Type}
\format{An object of class \code{identifier} of length 4.}
\usage{
xsd_string

rdfs_label
}
\description{
XSD String Type

Has Label Property
}
\keyword{datasets}
