% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsd-type.R
\docType{data}
\name{xsd_types}
\alias{xsd_types}
\alias{xsd_string}
\title{XSD Types}
\format{An object of class \code{identifier} of length 4.}
\usage{
xsd_string
}
\description{
\code{xsd_string} provides a string datatype
}
\keyword{datasets}
