% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{xsd_uri}
\alias{xsd_uri}
\title{XSD URI Type}
\format{An object of class \code{identifier} of length 4.}
\usage{
xsd_uri
}
\description{
XSD URI Type
}
\seealso{
Other semantic elements: \code{\link{blank_node}},
  \code{\link{rdf_type}}, \code{\link{rdfs_label}},
  \code{\link{semantic_elements}}, \code{\link{xsd_date}},
  \code{\link{xsd_integer}}, \code{\link{xsd_string}}
}
\concept{semantic elements}
\keyword{datasets}
