% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynvector.R
\docType{data}
\name{DynVector}
\alias{DynVector}
\title{An Amortized Vector}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DynVector
}
\description{
This object stores a preallocated list. You can add objects to the list
with \code{add} and add lists to the list with \code{add_list}. When the
objects to be added exceed the dimensions of the list, the list is re-
allocated with double the size. This amortizes the runtime if you keep
growing the dynamic vector.
}
\details{
Use this object if you need an object with a mutable states that you
keep growing. An example with be a list of RDF triples.
}
\examples{

v = DynVector$new(3)
v$add("test")

}
\keyword{datasets}
