% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier.R
\name{identifier}
\alias{identifier}
\title{Identifier Consturction}
\usage{
identifier(id, prefix = NA, blank = FALSE)
}
\arguments{
\item{id}{\code{character}. Local ID, for example a UUID. The part of
identifier after the prefix.}

\item{prefix}{named \code{character}. The name corresponds to the
prefix and the proper part to the namespace. Only the first element
of the vector will be honored. If you don't supply a prefix, the ID
will be treated as a URI and the QNAME and URI will be the same.}

\item{blank}{optional \code{logical}. If you want to create a blank node.}
}
\value{
\code{identifier} object (a type of list).
}
\description{
This is the constructor function for objects of the \code{identifier}
class.
}
\details{
An identifier in the semantic web is something that uniquely identifies
a resource. Identifiers can be represented as URI's (e.g.
\code{<http://example.com/id>}), or as QNAME's (e.g. \code{example:id}).

The Semantic Web model also allows for resources to be anonymous, via
so-called blank nodes. We use identifiers whose QNAME prefix is an
underscore (e.g. \code{_:alice}).

RDF4R stores identifiers as lists with the following fields:

\code{sample_id = list(
  id  = "57d68e07-8315-4b30-9a8e-57226fd815d7",
  uri = "<http://openbiodiv.net/57d68e07-8315-4b30-9a8e-57226fd815d7>",
  qname = "openbiodiv:id",
  prefix = c(openbiodiv = "http://openbiodiv.net")
)}
}
\examples{

a = identifier(
  id = "57d68e07-8315-4b30-9a8e-57226fd815d7",
  prefix = c(openbiodiv = "http://openbiodiv.net")
)

b = identifier(
  id = "alice",
  blank = TRUE
)

a
b

}
\seealso{
Other identifier functions: \code{\link{is.identifier}},
  \code{\link{print.identifier}},
  \code{\link{represent.identifier}}
}
\concept{identifier functions}
