% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier.R
\name{identifier}
\alias{identifier}
\alias{fidentifier}
\title{Identifier Consturction}
\usage{
identifier(id, prefix = NA)

fidentifier(label, ..., FUN = list(...), prefixes, def_prefix)
}
\arguments{
\item{id}{\code{character} Local ID. E.g., a UUID. The part of
identifier after the prefix.}

\item{prefix}{Named \code{character}. The first one will be used. If
there is no prefix supplied QNAME and URI will essentially be the
same.}

\item{label}{\code{character(1)} Parameter that will be passed to the lookup
functions. See \code{...} for details.}

\item{...}{(lookup) functions that will be executed in order to
 obtain the identifier. The functions should have one argument to which
 \code{label} will be assigned during the call. As soon as we have a
 unique match the function execution halts. If there is no match, a
 URI with the base prefix (the one indiciated by "_base") and a UUID
 will be generated.
@param FUN list of lookup functions to be tried. this can be omitted and
 instead the functions specified as additional arguments.}

\item{prefixes}{Named \code{character}. Contains the prefixes.}

\item{def_prefix}{The prefix to be used if lookup fails.}
}
\value{
\code{identifier} object (a type of list).
}
\description{
An identifier in the semantic web is something that uniquely identifies
a resource. Identifiers can be represented as URI's (e.g.
\code{<http://example.com/id>}), or as QNAME's (e.g. \code{example:id}).
}
\details{
RDF4R stores identifiers as lists with the following fields:

\code{sample_id = list(
  id  = "57d68e07-8315-4b30-9a8e-57226fd815d7",
  uri = "<http://openbiodiv.net/57d68e07-8315-4b30-9a8e-57226fd815d7>",
  qname = "openbiodiv:id",
  prefix = c(openbiodiv = "http://openbiodiv.net")
)}
}
\section{Functions}{
\itemize{
\item \code{identifier}: constructs an unique identifier from a local id and a prefix.

\item \code{fidentifier}: construct identifier via a lookup function.
}}

\examples{

}
