% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier.R
\name{identifier_factory}
\alias{identifier_factory}
\title{Manufacturing Identifier Constructors}
\usage{
identifier_factory(fun, prefixes, def_prefix)
}
\arguments{
\item{fun}{\code{list} of lookup functions to be tried. The functions
should have the same arguments (see \code{...} below). As soon as we have
a unique match the function execution halts. If there is no match, a
URI with the base prefix (the one indiciated by "_base") and a UUID
will be generated.}

\item{prefixes}{Named \code{character} that contains the prefixes to be
used for identifier construction after each lookup.}

\item{def_prefix}{to use if all else fails}
}
\value{
a function. The first argument of the function is a list. Each list
  element is a list of arguments to be passed to the lookup function.
}
\description{
Manufacturing Identifier Constructors
}
\examples{

openbiodiv_id = identifier_factory(simple_lookup,
  prefixes = prefixes,
  def_prefix = c(openbiodiv = "http://openbiodiv.net/"))

openbiodiv_id("Teodor Georgiev")
openbiodiv_id("Pavel Stoev")
}
