% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pasteif}
\alias{pasteif}
\title{Paste If}
\usage{
pasteif(..., sep = "", cond, return_value = NULL)
}
\arguments{
\item{...}{one or more R objects, to be passed to paste0 if cond is TRUE.}

\item{sep}{default "" - equiv. to paste0}

\item{cond}{boolean, condition to be true.}

\item{return_value}{}
}
\value{
string. If condition is false returns \code{return_value}. default is NULL
}
\description{
Pastes something only if the condition is met.
}
\examples{
pasteif("1", "st", cond = (3 < 2))

}
