% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier.R
\name{represent.identifier}
\alias{represent.identifier}
\title{Outputs an identifier in a default way (not print)}
\usage{
\method{represent}{identifier}(id)
}
\arguments{
\item{id}{\code{identifier}}
}
\value{
\code{character} default representation.
}
\description{
Outputs an identifier in a default way (not print)
}
\seealso{
Other identifier functions: 
\code{\link{identifier}()},
\code{\link{is.identifier}()},
\code{\link{print.identifier}()}

Other representables: 
\code{\link{represent.literal}()},
\code{\link{represent}()}
}
\concept{identifier functions}
\concept{representables}
