% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literal.R
\name{represent.literal}
\alias{represent.literal}
\title{Outputs a literal in a default way (not print!)}
\usage{
\method{represent}{literal}(ll)
}
\arguments{
\item{ll}{\code{literal}}
}
\value{
\code{character} default representation.
}
\description{
Outputs a literal in a default way (not print!)
}
