% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{xsd_string}
\alias{xsd_string}
\alias{rdfs_label}
\title{XSD String Type}
\format{An object of class \code{identifier} of length 4.}
\usage{
xsd_string

rdfs_label
}
\description{
XSD String Type

Has Label Property
}
\seealso{
Other semantic elements: \code{\link{blank_node}},
  \code{\link{rdf_type}}, \code{\link{semantic_elements}},
  \code{\link{xsd_date}}, \code{\link{xsd_integer}}

Other semantic elements: \code{\link{blank_node}},
  \code{\link{rdf_type}}, \code{\link{semantic_elements}},
  \code{\link{xsd_date}}, \code{\link{xsd_integer}}
}
\concept{semantic elements}
\keyword{datasets}
